// src/services/rest/interface.ts
export abstract class Entity implements IEntity {
    id?: string;

    constructor(data?: IEntity) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
        }
    }

    static fromJS(data: any): Entity {
        data = typeof data === 'object' ? data : {};
        throw new Error("The abstract class 'Entity' cannot be instantiated.");
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        return data;
    }
}
// ...
export class AquariumItem extends Entity implements IAquariumItem {
    aquarium!: string;
    name!: string;
    species!: string;
    inserted?: string;
    amount!: number;
    description?: string;

    constructor(data?: IAquariumItem) {
        super(data);
    }

    init(_data?: any) {
        super.init(_data);
        if (_data) {
            this.aquarium = _data["aquarium"];
            this.name = _data["name"];
            this.species = _data["species"];
            this.inserted = _data["inserted"];
            this.amount = _data["amount"];
            this.description = _data["description"];
        }
    }

    static fromJS(data: any): AquariumItem {
        data = typeof data === 'object' ? data : {};
        let result = new AquariumItem();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["aquarium"] = this.aquarium;
        data["name"] = this.name;
        data["species"] = this.species;
        data["inserted"] = this.inserted;
        data["amount"] = this.amount;
        data["description"] = this.description;
        super.toJSON(data);
        return data;
    }
}

export interface IAquariumItem extends IEntity {
    aquarium: string;
    name: string;
    species: string;
    inserted?: string;
    amount: number;
    description?: string;
}

export class Coral extends AquariumItem implements ICoral {
    coralType?: CoralType;

    constructor(data?: ICoral) {
        super(data);
    }

    init(_data?: any) {
        super.init(_data);
        if (_data) {
            this.coralType = _data["coralType"];
        }
    }

    static fromJS(data: any): Coral {
        data = typeof data === 'object' ? data : {};
        let result = new Coral();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["coralType"] = this.coralType;
        super.toJSON(data);
        return data;
    }
}

export interface ICoral extends IAquariumItem {
    coralType?: CoralType;
}

export enum CoralType {
    SoftCoral = "SoftCoral",
    HardCoral = "HardCoral",
}